fix(dashboard): move execution-log helpers out of route file so production builds pass
next build fails on main with:

  Type error: Route "src/app/api/workflows/crons/[agent]/executions/route.ts"
  does not match the required types of a Next.js Route.
  "readExecutionLogPage" is not a valid Route export field.

Next.js only allows HTTP handlers and route-config exports from route
files; the agent-level executions route also exported the
readExecutionLogPage/entriesToCsv helpers for reuse by the per-cron
sibling route, which type-checks in dev but is rejected by production
builds.

Move the helpers, their types, and the VALID_AGENT matcher verbatim to
dashboard/src/lib/cron-execution-log.ts and import them in both
executions routes. Pure refactor, no behavior change.

Verified: next build --webpack compiles clean; the 26 tests in
dashboard/src/app/api/workflows/crons/__tests__ pass unmodified.
